add .env fix token from .env add helmet for header request add crypto-js

+++ controllers/sauces.js
@@ -31,28 +31,42 @@ exports.getOneSauces = (req, res, next) => {
 };
 
 exports.modifySauces = (req, res, next) => {
+    Sauces.findOne({ _id: req.params.id })
+        .then((sauce) => {
+            if (sauce.userId !== req.auth.userId) {
+                return res.status(403).json({ error: "Accès non autorisé" });
+            } else {
+                const saucesObject = req.file ? {
+                    ...JSON.parse(req.body.sauce),
+                    imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+                } : {...req.body };
+                Sauces.updateOne({ _id: req.params.id }, {...saucesObject, _id: req.params.id })
+                    .then(() => res.status(200).json({ message: 'Objet modifié !' }))
+                    .catch(error => res.status(400).json({ error }));
 
-    console.log({...req.body });
+            }
+        })
+        .catch(error => res.status(500).json({ error }));
 
-
-    const saucesObject = req.file ? {
-        ...JSON.parse(req.body.sauce),
-        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
-    } : {...req.body };
-    Sauces.updateOne({ _id: req.params.id }, {...saucesObject, _id: req.params.id })
-        .then(() => res.status(200).json({ message: 'Objet modifié !' }))
-        .catch(error => res.status(400).json({ error }));
 };
+
 
 exports.deleteSauces = (req, res, next) => {
     Sauces.findOne({ _id: req.params.id })
         .then(sauces => {
-            const filename = sauces.imageUrl.split('/images/')[1];
-            fs.unlink(`images/${filename}`, () => {
-                Sauces.deleteOne({ _id: req.params.id })
-                    .then(() => res.status(200).json({ message: 'Objet supprimé !' }))
-                    .catch(error => res.status(400).json({ error }));
-            });
+            if (sauces.userId !== req.auth.userId) {
+                return res.status(403).json({ error: "Accès non autorisé" });
+            } else {
+                const filename = sauces.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${filename}`, (err) => {
+                    if (err) {
+                        console.log('error deleting file', err);
+                    }
+                    Sauces.deleteOne({ _id: req.params.id })
+                        .then(() => res.status(200).json({ message: 'Objet supprimé !' }))
+                        .catch(error => res.status(400).json({ error }));
+                });
+            }
         })
         .catch(error => res.status(500).json({ error }));
 };
